Break out all determination of dialect into something common between cldb and run
